+++ nctest/rec.c
@@ -340,8 +340,8 @@ test_ncrecput(const char *path) /* name of writable netcdf file to open */
 
     /* get a block of data of the right type for each record variable */
     for (iv = 0; iv < nrvars; iv++) {
-	datap[iv] = emalloc(rvarsizes[iv]);
-	datar[iv] = emalloc(rvarsizes[iv]); /* for comparison values */
+	datap[iv] = emalloc((size_t)rvarsizes[iv]);
+	datar[iv] = emalloc((size_t)rvarsizes[iv]); /* for comparison values */
 	if (ncvarinq(ncid, rvarids[iv], 0, &vartype[iv], 0, 0, 0) == -1) {
 	    error("%s: ncvarinq failed", pname);
 	    ncclose(ncid);
@@ -497,8 +497,8 @@ test_ncrecget(const char *path) /* name of netcdf file to open */
 
     /* get a block of data of the right type for each record variable */
     for (iv = 0; iv < nrvars; iv++) {
-	datap[iv] = emalloc(rvarsizes[iv]);
-	datar[iv] = emalloc(rvarsizes[iv]); /* for comparison values */
+	datap[iv] = emalloc((size_t)rvarsizes[iv]);
+	datar[iv] = emalloc((size_t)rvarsizes[iv]); /* for comparison values */
 	if (ncvarinq(ncid, rvarids[iv], 0, &vartype[iv], 0, 0, 0) == -1) {
 	    error("%s: ncvarinq failed", pname);
 	    ncclose(ncid);
